+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 0.6.4.1 (Unreleased)
 
 * Making setupAction on ServiceCollection DI extension method optional
+* Updating signature for Sys.Renew
 
 ## 0.6.4